+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Added:
 - 
 ### Fixed:
-- 
+- Fixed config list
+- Fixed config list delimiter
 ### Changed:
 - Changed config from json5 to json